Simply tox Django/Wagtail factors into "current" and "next"
This change takes our matrixed Django and Wagtail factors in our tox configuration and simplifies them into "current" and "next". The intention is to make it easier for devs to select "the currently supported versions" versus "upcoming versions we need to support" without having to know or care what those versions are for the purposes of invoking tox.

It replaces the `dj{111}-wag{113}` part of the master environments with `{current,next}` factors, meaning invoking those environments directly is now:

```
tox -e unittest-py36-current-slow
```

for example.

It also adds a convenience `next` test environment that can be invoked:

```
tox -e next
```

For now it also adds Wagtail 2.3, but keeps Django 1.11 (2.3 is the last version of Wagtail to support Django 1.11).