Adds script to transfer enforcement actions to new page type
After https://github.com/cfpb/cfgov-refresh/pull/5523 we want
to move old DocumentDetail pages to our new EnforcementAction
page type. This adds a script that does that as well as a
small util that manages html tags potentially embedded in the
current pseudo-metadata fields.

Currently, this script doesn't split or manage the 3 extant
pages with multiple statuses, but that will be done manually
after this script is run.